+++ src/components/about-us/section-04/index.js
@@ -1,15 +1,16 @@
+import { colors } from '../../../themes/common-variables'
 import { font, marginBetweenSections } from '../constants/styles'
-import { replaceStorageUrlPrefix } from '@twreporter/react-components/lib/shared/utils'
+import { replaceGCSUrlOrigin } from '@twreporter/core/lib/utils/storage-url-processor'
 import { screen } from '../utils/screen'
 import { storageUrlPrefix } from '../utils/config'
+import MoreInfo from './more-info'
+import React, { PureComponent } from 'react'
+import VelocityComponent from '@twreporter/velocity-react/velocity-component'
 import chunk from 'lodash/chunk'
 import data from '../constants/section-04/partners'
 import groupBy from 'lodash/groupBy'
 import keys from 'lodash/keys'
-import MoreInfo from './more-info'
-import React, { PureComponent } from 'react'
 import styled from 'styled-components'
-import VelocityComponent from '@twreporter/velocity-react/velocity-component'
 
 const _ = {
   chunk, groupBy, keys
@@ -33,6 +34,7 @@ const containerWidth = {
 
 const Container = styled.div`
   position: relative;
+  background-color: ${colors.white};
   ${screen.overDesktop`
     margin: ${marginBetweenSections.overDesktop} 0;
   `}
@@ -73,7 +75,7 @@ const SectionWrapper = styled.section`
 `
 
 const Title = styled.h1`
-  background-image: url(${replaceStorageUrlPrefix(`${storageUrlPrefix}/title-section4.png`)});
+  background-image: url(${replaceGCSUrlOrigin(`${storageUrlPrefix}/title-section4.png`)});
   background-repeat: no-repeat;
   background-size: contain;
   margin: 0;
@@ -291,7 +293,7 @@ export default class Section4 extends PureComponent {
               onClick={() => this._select(index)}
             >
               <LogoContent>
-                <img src={replaceStorageUrlPrefix(data.logo)} />
+                <img src={replaceGCSUrlOrigin(data.logo)} />
                 <h3>{data.name.english}</h3>
                 <p>{data.name.chinese}</p>
               </LogoContent>
@@ -302,7 +304,7 @@ export default class Section4 extends PureComponent {
             onClick={() => this._select(index)}
           >
             <LogoContent>
-              <img src={replaceStorageUrlPrefix(data.logo)} />
+              <img src={replaceGCSUrlOrigin(data.logo)} />
               <div>
                 <h3>{data.name.english}</h3>
                 <p>{data.name.chinese}</p>