+++ src/components/article/Body.js
@@ -2,11 +2,11 @@
 'use strict'
 import PropTypes from 'prop-types'
 import React, { PureComponent } from 'react'
-import Waypoint from 'react-waypoint'
 import constStyledComponents from '../../constants/styled-components'
 import getArticleComponent from './getArticleComponent'
 import sideBarFactory from '../side-bar/side-bar-factory'
 import styled from 'styled-components'
+import { Waypoint } from 'react-waypoint'
 import { articleLayout } from '../../themes/layout'
 import { screen } from '../../themes/screen'
 import { typography } from '../../themes/common-variables'
@@ -26,15 +26,16 @@ const _ = {
 
 const StyledArticleComponent = styled(constStyledComponents.ResponsiveContainerForAritclePage)`
   font-size: ${(props) => {
-    let fontSize = typography.font.size.base
+    let fontSize
     switch(props.fontSize) {
-      case 'small':
-        fontSize = typography.font.size.xSmall
+      case 'medium':
+        fontSize = '20px'
         break
       case 'large':
-        fontSize = typography.font.size.larger
+        fontSize = '22px'
         break
       default:
+        fontSize = '18px'
         break
     }
     return fontSize