+++ src/constants/index.js
@@ -1,6 +1,5 @@
 'use strict'
 
-export * from './action-types'
 export * from './footer'
 export * from './link-prefix'
 export * from './site-meta'
@@ -75,7 +74,7 @@ export const basePath = 'https://www.twreporter.org'
 
 export const appId = 962589903815787
 
-export const donatePath = 'https://twreporter.org/donation/period'
+export const donatePath = 'https://twreporter.org/donation'
 
 export const categoryPath = {
   taiwanPath: '/category/taiwan',