+++ src/express/middlewares/init-redux-store.js
@@ -1,5 +1,10 @@
-import configureStore from '../../store/configure-store'
+/* global __DEVELOPMENT__ */
+import requestOrigins from '../../constants/request-origins'
+import twreporterRedux from '@twreporter/redux'
+// lodash
 import set from 'lodash/set'
+
+const stateFields = twreporterRedux.reduxStateFields
 
 const _ = {
   set
@@ -9,12 +14,19 @@ const _ = {
  *  This middleware inits redux store,
  *  and set it as `req[namespace].reduxStore`
  *  @param {string} namespace - namespace is used in `req` object to avoid from overwriting the existed field
+ *  @param {Object} options
+ *  @param {string} options.releaseBranch - release branch, it could be 'master', 'test', 'staging' or 'release'
  */
-function initReduxStoreMiddleware(namespace) {
+function initReduxStoreMiddleware(namespace, options) {
   return async function middleware(req, res, next) {
     try {
       const cookieList = req.get('cookie')
-      const reduxStore = await configureStore({}, cookieList)
+      // The redux actions will take the `origins` in the store, and use them as the origins of request url.
+      // We set the initial origins for server side renedering. And we'll set them for client side rendering before we send the store to client.
+      const initState = {
+        [stateFields.origins]: requestOrigins.forServerSideRendering[options.releaseBranch]
+      }
+      const reduxStore = await twreporterRedux.createStore(initState, cookieList, __DEVELOPMENT__)
       _.set(req, [ namespace, 'reduxStore' ], reduxStore)
       next()
     } catch(err) {