+++ src/managers/layout-manager.js
@@ -1,12 +1,12 @@
 import Footer from '@twreporter/react-components/lib/footer'
 import React from 'react'
 import styled from 'styled-components'
-import uh from '@twreporter/universal-header'
+import Header from '@twreporter/universal-header/dist/containers/header'
 import { colors } from '../themes/common-variables'
 import { screen } from '../themes/screen'
 import { themesConst } from './theme-manager'
 
-const HeaderContainerWithTransparentTheme = styled.div`
+const FullScreenHeader = styled.div`
   position: absolute;
   top: 0;
   left: 0;
@@ -15,6 +15,11 @@ const HeaderContainerWithTransparentTheme = styled.div`
   ${screen.mobile`
     position: relative;
   `}
+`
+
+const PinkBackgroundHeader = styled.div`
+  position: relative;
+  background-color: #fabcf0;
 `
 
 const styles = {
@@ -68,16 +73,36 @@ export default class LayoutManager {
       case themesConst.withoutHeaderAndFooter: {
         return null
       }
-      case themesConst.articlePage.fullscreen.dark:
-      case themesConst.articlePage.fullscreen.normal: {
+      case themesConst.articlePage.v2.pink: {
         return (
-          <HeaderContainerWithTransparentTheme>
-            <uh.Header
+          <PinkBackgroundHeader>
+            <Header
               theme="transparent"
               releaseBranch={this.releaseBranch}
               isLinkExternal={false}
             />
-          </HeaderContainerWithTransparentTheme>
+          </PinkBackgroundHeader>
+        )
+      }
+      case themesConst.articlePage.v2.photo: {
+        return (
+          <Header
+            theme={themesConst.photography}
+            releaseBranch={this.releaseBranch}
+            isLinkExternal={false}
+          />
+        )
+      }
+      case themesConst.articlePage.fullscreen.dark:
+      case themesConst.articlePage.fullscreen.normal: {
+        return (
+          <FullScreenHeader>
+            <Header
+              theme="transparent"
+              releaseBranch={this.releaseBranch}
+              isLinkExternal={false}
+            />
+          </FullScreenHeader>
         )
       }
       // TODO Header of topic page should be implmeneted here, rather than in topicLandingPage container
@@ -86,7 +111,7 @@ export default class LayoutManager {
       //}
       default: {
         return (
-          <uh.Header
+          <Header
             theme={this.theme}
             releaseBranch={this.releaseBranch}
             isLinkExternal={false}